Refactor: Extract interaction handling to InteractionHandler
Moved the logic for handling interactions in the DishDetailsViewModel to a new, stateless `InteractionHandler` class.

This change aims to improve code organization and testability by separating the interaction handling concern from the ViewModel.
